requirements.txt
django
python-dotenv
psycopg2-binary
requests
django-cors-headers
redis
gunicorn
pytz
